[WIP] CUAEV Double Precision (#181)
* CUAEV Double Precision

* launch_cuAngularAEVs_kernel

* docs tensorboard

* cuRadialAEVs

* cuaev_forward nbrlist launch function

* finish backward or doublebackward refactor

* template to support both float and double precision

* Rcr double is 5.1, float is 5.099999904632568

* unittest

* quick save

+++ torchani/csrc/aev.h
@@ -125,8 +125,8 @@ struct NeighborList {
 };
 
 struct AEVScalarParams {
-  float Rcr;
-  float Rca;
+  double Rcr;
+  double Rca;
   int radial_sublength;
   int radial_length;
   int angular_sublength;
@@ -141,8 +141,8 @@ struct AEVScalarParams {
   bool use_cos_cutoff;
 
   AEVScalarParams(
-      float Rcr,
-      float Rca,
+      double Rcr,
+      double Rca,
       Tensor EtaR_t,
       Tensor ShfR_t,
       Tensor EtaA_t,
